requirements.txt
pandas
scipy
numpy
geopy
termcolor
matplotlib
shapely
pyyaml
IPython
pprint
